Refactoring and moving methods under query manager
Methods moved as they can be used to query different resources
Updates to the unit tests where needed
